Классы с отдельными фреймами для каждого дерева заменены на один класс с общим фреймом. Таким образом не нужно будет менять каждый раз компонент
