--- App/src/main/kotlin/guiClasses/components/nodePanels/AVLPanel.kt
@@ -1,10 +1,0 @@
-package guiClasses.components.nodePanels
-
-import guiControl.painters.AVLPainter
-import nodes.AVLNode
-import trees.AVLTree
-
-class AVLPanel(private val tree: AVLTree<Int, Int>) :
-    AbstractNodePanel<AVLNode<Int, Int>, AVLTree<Int, Int>, AVLPainter>(tree) {
-    override fun painterInit() = AVLPainter(tree, Constants.nodeMargin, Constants.nodeSize, width)
-}
